Install the node libs (without noderfc) globally
node-rfc is required - of course - but we cannot install
it globally. Reason seems to be (not sure about that): inside
node-rfc there is another npm install which gets not forwarded
the global option.

Installing the libs locally into the current project folder does
not work in conjunction with the Jenkins-docker-plugin since in this
case the plugin controls the working dir (-w). Caused hereby locally
installed node libs are not found anymore.