Ensure a simple --version command does not fail
If mtaBuild --version was called without mounting anything into the container we failed since .mta could not be
removed. The underlying cause was a permission issue. This is now solved by chmod to 777. In general - as we
all know - this is not a cool approach. But "in the wild" we expect a project directory mounted from outside
anyway.